moved AccessAvatars component into JourneyCard, removed avatars config
